Use PID files to stop exec/opensnoop processes
The current demo commands were run for 10 seconds and
could not be stopped earlier. A way to let the snoopers
run until Ctrl-C is pressed is needed.

Since systemd-run is not available, simple shell commands
are used to create a PID file and later kill the process.
Maybe if we migrate to go-client instead of forking local
"kubectl exec" commands there is a nicer inbuilt way.