Use client-go instead of spawning kubectl
Forking kubectl processes with the same stdout was an
easy way to run processes on the pod.
Now client-go is used for that with the advantage to
run in go routines that are terminated at program exit
and don't write to stdout any further when the shell
prompt is already there.