Add first iteration of the Gadget Tracer Manager
The goal is to avoid string comparison in BPF, as well as enabling
future features with uprobes.

Problem statement
=================

We want to filter pods by labels (`kubectl gadget execsnoop --label
role=demo`). Although the initial version of inspektor-gadget managed to
do that, it suffers from the following limitations:

- the string comparisons is done in the BPF program and that's an
expensive operation, specially when it is in the hot path (kprobe on a
kernel function that is called often)

- we don't have real strcmp() functions in BPF. The initial version
limits the label string to 64 characters. If the label is longer, not
all characters will be checked, leading to an incorrect result. If the
string is shorter, the program will still check 64 characters, making it
more expensive than necessary.

- the program is complex to write: the C program is generated from a
python script. While it worked fine for one label check (`--label
role=demo`), it makes it more complicated to add full label filters
(`k8s-app=myapp,role=demo`) or add additional filters on the container
images (`docker.io/albanc/toolbox:latest`), annotations, the command or
anything else in the pod yaml. This logic is also copy/pasted in several
bcc scripts, without a good way to factorise the logic.

Proposed solution
=================

We have a new daemon set called "gadget tracer manager" implementing a
gRPC interface and updating BPF hash maps.

Each deamon instance on each node is independent: they don't need to
communicate between themselves and don't need coordination. They only
manage traces on the node they are running on. This makes it simple.

Methods called by the bcc script:
- AddTracer(labelSelector, imageSelector...) -> (tracer_id string, err error)
  - Called whenever a new tracer is started (i.e. `kubectl gadget execsnoop...`)
  - Create a new instance of BPF hash map, described below
- RemoveTracer(tracer_id string) -> (err error)
  - Called whenever a tracer is stopped (i.e. `^C`)
  - Unpin the instance of BPF map

The gRPC client here is running in the same pod as the daemon.

Methods called by the OCI hooks:
- AddContainer(cgroup uint64, containerDefinition) -> (err error)
- RemoveContainer(cgroup uint64) -> (err error)

The gRPC client here is running on the host. The socket file therefore
needs to be in a directory that is shared with the host (e.g.
/run/gadgettracermanager.socket).

BPF Hash map
============

Each tracer has its own instance of the following BPF hash map:
- pinned path: /sys/fs/bpf/gadget/cgroupidset-a8e5f7d217b8
- key: cgroup u64
- value: dummy

The manager will add the cgroup in the map instance:
- when the map is created via the AddTracer method, it needs to be
  initialized with the cgroups already existing that match the label
  selector.
- whenever there is a new pod matching the criteria of the tracer,
  during the OCI Prestart hook, via the AddContainer() method.

BPF Program
===========

The pseudo code for the BPF program is a lot simpler:
```
u64 cgroupid = bpf_get_current_cgroup_id();
exists = tracer_a8e5f7d217b8.lookup(&cgroupid)
if (!exists)
  return 0;
```

In this way, all the logic about the labelSelector is moved to userspace.

BCC script
==========

Instead of adding Kubernetes-specific parameters like `--label`,
`--podname` etc. that are unlikely to be accepted in bcc upstream, we
could add a single generic parameter specifying the BPF hash map to
filter cgroups, populated and managed externally:

```
/usr/share/bcc/tools/opensnoop --cgroup-custom-filter=/sys/fs/bpf/gadget/cgroupidset-a8e5f7d217b8
```

inspektor-gadget would not kubectl-exec opensnoop directly but use a
wrapper script that will:

- call the AddFilter() method
- run the opensnoop script with `--cgroup-custom-filter=` as specified
by the return value of AddFilter()
- call the RemoveFilter() method