Introduce flag to disable traceloop
Introduce the traceloop flag, If false, do not start the traceloop daemon.
This is useful to avoid the memory consumption of the traceloop gadget, if only
the other gadgets are desired.

kubectl gadget traceloop commands return an error if traceloop is not enabled.

  kubectl gadget deploy <options>
      --traceloop           enable the traceloop gadget (default true)

This will set the environment variable $INSPEKTOR_GADGET_OPTION_TRACELOOP in
the gadget DaemonSet. The script entrypoint.sh will then act on it.