cmd/inspektor-gadget: Fix terminal output
The postProcessor is a class that wraps an io.Writer object to add a prefix
to all the lines that are printed to the terminal, it's used to print the
node number before each line.

The current logic introduces in some cases a line break making it difficult to
read. The following is a snippet of the opensnoop gadget:

...
[ 0] 19651  sh                  3   0 /etc/passwd
[ 0] 19651  sh                  3   0
[ 0] /etc/passwd
...

This problem happens because the Write() function doesn't receive a full string
in all the cases, for instance in the last example it received something like:

Write("19651  sh                  3   0 /etc/passwd\n")
Write("19651  sh                  3   0")
Write("/etc/passwd\n")

This commit fixes the logic by reworking the logic to post process the output
from the BCC scripts.