networkpolicyadvisor: batch events to reduce api calls
The Network Policy Advisor receives TCP connection tuples with IP
addresses from the BPF program. In order to match them with Kubernetes
pods and services, it needs to make two API calls to the kube-apiserver
to get the IP of the pods and the IP of the services.

Before this patch, the kube-apiserver was called for each new TCP
connection. This takes too much CPU and it is sometimes unable to handle
the connections fast enough.

With this patch, it uses a ticker and it can only call the API server
maximum every second and it caches the result for a batch of TCP
connections. It takes care of making the API calls after receiving the
TCP connection to avoid the race of evaluating a connection from a new
pod based on a old cache of the pod IPs.