gadgets/networkpolicyadvisor: Use k8sutil to create k8s client set
The network policy advisor running inside the cluster used
`clientcmd.BuildConfigFromFlags` to create a configuration. This function
prints a warning "Neither --kubeconfig nor --master was specified.  Using the
inClusterConfig.  This might not work." if kubeconfig is "".

This commit changes the network policy advisor to use NewClient function from
the github.com/kinvolk/inspektor-gadget/pkg/k8sutil and updates this function
to use rest.InClusterConfig() when kubeconfig is "".