ocihookgadget: Remove requirement to have a kubeconfig
ocihookgadget is invoked each time a container is created or deleted.
The information received from the OCI hooks is not enough and the k8s API
server is used to get the pod name, namespace, labels and container index.
ocihookgadget runs in the host, hence it needs a valid kubeconfig to contact
the API server. Before this commit, it used the same kubeconfig as the kubelet
running in the node. Even if this approach worked because the permissions
needed by the kubelet are enough, it was not clean. This commit moves the logic
to fetch the information from the k8s API server to the gadgettracer that
runs inside the gadget container and can use the in cluster configuration to
access the API server.