Add access to /usr/src without direct bind-mount
Ubuntu has /usr/src with the kernel headers but on Flatcar Container
Linux, this directory does not exist and the kernel headers are located
in /lib/modules/$(uname -r) directly.

Attempting to mount /usr/src from the host would cause Docker to attempt
to create that directory but this does not work when the host /usr is
read-only.

This patch adds access to /usr/src with a symbolic link, so it works
both for Linux distributions with and without /usr/src.