Network policy advisor: Align "--namespace" and "--namespaces" flags
The network policy advisor gadget has a --namespaces (notice the "s") to specify
the namespaces to consider when monitoring traffic. This option was isolated
from the "--namespace" flag, so a command like "--namespace demo" will monitor
traffic in the "default" namespace (default value for "--namespaces").

This commit adds a simple logic to assign "--namespace" to "--namespaces" if
"--namespaces" is not specified, this avoids the confusion.