pkg/gadgettracermanager: remove bpf file and assets
The gadget tracer manager creates (and pins) an eBPF map that is used
by the gadget to filter the events of the different containers.

Before this commit the maps were defined in a .c file that was compiled
with clang/llvm into a .o file, then that file was embedded into a
.go file by using go-bindata. This process is so complex and not needed
because those maps can be created (and pinned) just performing some
system calls.

This commit uses the cilium/ebpf library to perform that task.

Signed-off-by: Mauricio Vásquez <mauricio@kinvolk.io>