Update build environment and dependencies
- Update cilium/ebpf to add support for BPF iterators.
- Update Ubuntu to Focal LTS (20.04) so we can have libbpf-dev.
  libbpf-dev is available on Ubuntu Focal from a PPA. Note that we don't
  switch to the latest Ubuntu because we use the bcc container image that
  currently builds on 18.04 and 20.04.
- Update Go to 1.16 so we can have embed.
- Update gobpf so it works with go modules.
- Update tcptracer-bpf (with unmerged PR 69) so it works with go
  modules.

Additionally, update the Makefile minikube-install target so it does not
require an external registry.

In the developer workflow, it is cumbersome to push the gadget container
image to a container registry for every change, especially with bad
internet connectivity.

This change uses the 'minikube cache add' command.

Additionally, remove gadget-local.Dockerfile: this can cause problem if
the glibc versions on the compiler's host and in the gadget container
have different versions:

```
/bin/gadgettracermanager: /lib/x86_64-linux-gnu/libc.so.6: version `GLIBC_2.32' not found (required by /bin/gadgettracermanager)
```

Finally, add an option to disable the liveness probe to help debugging
gadgettracermanager crashes.