Makefile: minikube-install target does not require external registry
In the developer workflow, it is cumbersome to push the gadget container
image to a container registry for every change, especially with bad
internet connectivity.

This patch changes that by pushing the gadget container image to
Minikube's docker without uploading & downloading the image.

Additionally, remove gadget-local.Dockerfile: this can cause problem if
the glibc versions on the compiler's host and in the gadget container
have different versions:

```
/bin/gadgettracermanager: /lib/x86_64-linux-gnu/libc.so.6: version `GLIBC_2.32' not found (required by /bin/gadgettracermanager)
```

Finally, add an option to disable the liveness probe to help debugging
gadgettracermanager crashes.