Gadget Tracer Manager: export container details in BPF map
Before this patch, Gadget Tracer Manager creates BPF maps for each
tracer so that gadgets can filter events for specific containers.
However, that does not allow gadgets to display the container name for
each event.

This patch adds a new hash map "containers" exposing container details
for each mount namespace:

- Keys: mount namespace inode
- Values: Container ID, Kubernetes namespace, pod and container names
- Pin: /sys/fs/bpf/gadget/containers

This makes it possible for gadgets to access that information and
display it. Gadgets like opensnoop and execsnoop could later be
reimplemented to use this. External tools such as tracee or bpftrace
could also benefit from this just by using the "containers" map (other
interaction with Inspektor Gadget is not necessary for this).