entrypoint: remount bpf_fs as workaround
In correct configurations, the host has the following mountpoints:
- /sys of type sysfs, mounted in rw mode
- /sys/fs/bpf of type bpf_fs, mounted in rw mode

Then, the gadget pod is bind-mounting the host's /sys/fs/bpf into the
container's /sys/fs/bpf.

However, it does not work well with Minikube when using the Docker
driver: Minikube simulates the host with a container (the "kikbase"
container with the image gcr.io/k8s-minikube/kicbase:v0.0.10) running
systemd and the docker daemon. And the Kubernetes pods are created with
"inner containers" started by this "docker in docker". The problem here
is that the kicbase container has a read-only /sys and the bpf_fs is not
mounted at all.

The Gadget container does not have the control on the way the host is
configured. This patch merely detects the situation and remount a new
/sys/fs/bpf as a workaround.