CRI: return a better error when pid is zero
Sometimes, the container runtime says the pid of the container is zero:
```
Skip pod kube-system/gadget-prm95: cannot find cgroup path: cannot parse cgroup: open /proc/0/cgroup: no such file or directory
```

In this case, don't try to read /proc/0/cgroup but return an error
message early.

The OCI Container Runtime specification says that the pid is required
when the state of the container is 'created' or 'running' but not when
it is 'creating' or 'stopped'.