GitHub action: cache Go modules
Using the technique from the documentation:
https://github.com/actions/cache/blob/main/examples.md#go---modules

This is useful for building the kubectl plugin ("make kubectl-gadget")
and the unit tests ("make test"): those are built directly on the Ubuntu
host.

Before this patch:
			Test 1		Test 2
- make kubectl-gadget:	2m 42s		2m 18s
- make test:		57s		48s

After this patch, in case of cache hit:
- restoring the cache:	8s
- make kubectl-gadget:	12s
- make test:		3s

Note: this patch is complementary with the previous commit 93f2f767688e
("Dockerfile: faster builds") where we cache the Go modules in a Docker
layer. This was useful for building the gadget container image in Docker
("make gadget-container").