gadget-container/Makefile: prepare to compile BPF code
Future gadgets will have BPF code in the pkg/gadgets/.../bpf/
directories that need to be compiled with clang. This patch updates
gadget-container/Makefile so that the makefiles in such directories can
be called. In this way, gadget-container/Makefile does not have to be
updated each time we change something in pkg/gadgets/.../bpf/. This will
make it easier to consume Go packages outside of Inspektor Gadget.