New Resolver interface for gadgets to call back the GadgetTracerManager
Before this patch, the process gadgettracermanager started two main
independent components:
1. GadgetTracerManager, the gRPC server
2. TraceReconciler, a Controller Manager for the Trace custom resource

The TraceReconciler calls methods on TraceFactory and Trace interfaces.
But sometimes, a Trace needs to do a lookup on the GadgetTracerManager
to match a mntns and a container.

This patch passes a GadgetTracerManager reference to the
Trace.Operation() method, so it can do lookups. The reference is not
passed directly but via a Resolver interface, so it limits the API
exposed to gadgets.