kubectl-gadget: add --json flag
Add this flag to be passed to the gadgets that support it.
The postProcess logic is modified a bit to avoid removing the first
line when json is used.

The gadgets will output each event on a different line, therefore the
existing logic to buffer and print whole lines coming from the different
nodes will continue to work well on this case.