Trace and TraceFactory interfaces: add k8s client and capabilities
Gadgets implementing the Trace and TraceFactory interfaces now have
access to a k8s client. If they want to support a specific scheme, their
TraceFactory can implement the optional method AddToScheme().

This will be useful for the seccomp gadget. Other gadgets can ignore the
client argument and don't have to implement the AddToScheme() method.

The TraceFactory interface also has an optional method
SupportsOutputMode() so the controller will know which values outside of
OutputMode=Status are implemented by the gadget.