socket-collector: Add new gadget to collect TCP/UDP sockets information
This commit introduces a new gadget named socket-collector which, using BPF
iterators, provides a snapshot of all TCP and UDP sockets.

TCP/UDP BPF Iterators read information from the current network namespace.
Therefore, to apply filters on namespace or pod, we need to change
network namespace before opening the iterator. To do that, we use the
container PID retrieved with the function LookupPIDByPod and then we
use the netns package to change namespace.

Currently, it is necessary to specify both namespace and podname using
filters as prepared by the Gadget Tracer Manager. For instance:

apiVersion: gadget.kinvolk.io/v1alpha1
kind: Trace
metadata:
  name: socket-collector
  namespace: gadget
spec:
  node: my-node
  gadget: socket-collector
  filter:
    namespace: default
    podname: myapp1-pod-wsgx9
  runMode: Manual
  outputMode: Status

How to use:

$ kubectl apply -f  pkg/resources/samples/trace-socket-collector.yaml
$ kubectl annotate -n gadget trace/socket-collector gadget.kinvolk.io/operation=start
$ kubectl get -n gadget trace/socket-collector -o jsonpath='{.status.output}'
[
 {
  "protocol": "TCP",
  "local_address": "0.0.0.0",
  "local_port": 27017,
  "remote_address": "0.0.0.0",
  "remote_port": 0,
  "status": "LISTEN",
  "kubernetes_namespace": "default",
  "kubernetes_pod": "myapp1-pod-wsgx9"
 },
 {
  "protocol": "TCP",
  "local_address": "192.168.1.105",
  "local_port": 47158,
  "remote_address": "80.249.99.148",
  "remote_port": 80,
  "status": "ESTABLISHED",
  "kubernetes_namespace": "default",
  "kubernetes_pod": "myapp1-pod-wsgx9"
 },
 {
  "protocol": "UDP",
  "local_address": "192.168.1.105",
  "local_port": 68,
  "remote_address": "192.168.1.1",
  "remote_port": 67,
  "status": "ACTIVE",
  "kubernetes_namespace": "default",
  "kubernetes_pod": "myapp1-pod-wsgx9"
 }
]

In case of errors, status.output will be empty and error will be
described in status.operationError which can be read like this:

$ kubectl get -n gadget trace/socket-collector -o
jsonpath='{.status.operationError}'