gadgettracermanager: fix mutex usage
The methods to add/remove containers and tracers can be called
concurrently from different threads and they read/update the
'containers' and 'tracers' maps respectively. This patch makes use of
the existing mutex to protect them.

Additionally, the field containerIDsByKey is removed from
GadgetTracerManager and instead moved as a local variable to run().
Since it is only used from the same thread, it does not need to be
protected by the mutex.