New command: local-gadget
This patch makes the gadgets from Inspektor Gadget available on a Linux
machine without Kubernetes:
- kubectl-gadget: Inspektor Gadget on Kubernetes
- local-gadget: Inspektor Gadget on a single Linux machine

local-gadget can be executed without parameters and it will start a
shell with auto completion. Type "help" to get the list of commands.
Commands are implemented with cobra.

local-gadget uses the runc fanotify package to know when containers
based on runc are started or stopped. Notably it works with docker
containers.

Possible use cases:

- Faster development cycle of gadgets: they no longer need to be
  deployed on Kubernetes and the compilation of a single binary is
  faster than the build of the gadget container.

- Easier to write gadget tests: we can test whether a BPF program is
  accepted by the verifier without deploying to Kubernetes.

- Some gadgets have meaningful use cases on their own on a single Linux
  machine. E.g. generate seccomp profiles with Docker.

Compile with:
```
make local-gadget
```

Examples in docs/local-gadget.md