socket-collector: Make Socket-Event type portable
By defining the Socket-Event in a separate file, there will not be
dependencies problems when importing it on kubectl client side.

In addition, from this commit on, the socket-collector gadget starts
using the new common Event data structure defined in pkg/types to
avoid redefining the gadgets common fields, e.g kubernetes
information.