dns gadget: rework filter on networkHost pods
Don't check network namespace (/proc/pid/ns/net) of terminated
containers.

The tracer keeps a reference count when it is attached on a network
namespace. gadget.go might attempt to attach the tracer on the same
netns several times because:
1. a pod might have several containers
2. a pod might run with networkHost=true

In those cases, the tracer is added only one time, and it is detached
when the reference counter reaches zero.