Allow forks to test all GitHub Actions
When running a release, the GitHub Action "Push Integration Test Image"
pushes the container image docker.io/kinvolk/gadgettest.

Before this patch, the name was hardcoded, making it difficult to test
the GitHub Action in forks. With this patch, it uses $CONTAINER_REPO
with the "test" suffix, so forks don't need write access to
docker.io/kinvolk/gadgettest.