seccomp: fix fd garbage collector issue
Keep a reference on the eBPF link, even if we don't access it.
Otherwise, the garbage collector will free it and call the finalizer set
up by the cilium/ebpf library:

runtime.SetFinalizer(fd, (*FD).Close)

Symptoms of the bug before this fix: the seccomp gadget works fine but
stops working after a while (when the garbage collector decides to call
the finalizer and close the file descriptor).