runcfanotify: fix fd double close
The go-fanotify package has an inconsistent API: data.Close() and
data.File() cannot be used together because one is closing the file with
unix.Close() and the other is using os.NewFile() so it will be closed by
the os.File's finalizer once the garbage collector frees it.

Since we need the os.File to be able to use ioutil.ReadAll(), let's be
consistent and never call data.Close().