socket-collector: Update Go skeleton generated by bpf2go
The latest update of cilium/ebpf modifies the skeleton generated by bpf2go.

Now that the skeleton is pulling in the BPF elfs using the embed package
instead of a literal string, we need to commit also the generated .o to
avoid problems when running unit-tests and generating the documentation.