GitHub Action: check for any uncommitted changes
Check for any uncommitted changes and fail the GitHub Action if any is
found.

This is to avoid mistakes like in the previous release where a version
with the "-dirty" suffix was reported:
```
$ ./kubectl-gadget version
v0.3.0-dirty
```

This was because "bin/" was forgotten from .gitignore. With this patch,
we will no longer forget entries in .gitignore.