runcfanotify: notify termination of initial containers
runcfanotify uses a pidfd to be notified when a container terminates and
emit the container event EVENT_TYPE_REMOVE_CONTAINER. But this was only
done for containers that were created while RuncNotifier was running.

This patch adds a method AddWatchContainerTermination() to be able to
monitor containers that were started before RuncNotifier is started.

This can be called for initial containers detected by PodInformer, or
later by Docker in local-gadget.