Makefile: minikube-install: wait for uninstall before new installation
I have race conditions where the gadget namespace is not fully deleted
after the "kubectl delete -f". So the "kubectl apply" does not need to
create the namespace, but then the deletion terminates soon after.

Similarly, there are sometimes two gadgets pod: one in the terminating
state, and one in the creating state.

Additionally, it seems that "kubectl delete --wait" does not always
work. So I am adding an explicit "kubectl wait".