cmd: Implement columns parameter
By default the different gadgets print a lot of columns with information
that could be useless in some cases. For instance node name, namespace,
pod name, etc when the filter only matches one of them. This commit
implements a new column parameter so the user can choose which columns
to print and their order.