runcfanotify: AddWatchContainerTermination returns error
This makes AddWatchContainerTermination() return an error synchronously if
pidfd_open fails. This can happen if the container process is killed
quickly before being started.

This means that when ContainerCollectionInitialize() returns, the pidfds
are already opened instead of being opened asynchronously in the go
routine. This makes tests for fd leaks easier to write.