kubectl-gadget: Do not wait for timeout if trace updated the Status
After the Reconcile updates the Trace.Status either moving the state
forward or notifying an error/warning, such Status will never change
anymore because the Reconcile will remove the annotation regardless of
whether the operation requested in the annotation succeeded or not.
Meaning that a given operation will always be executed only once
unless the user adds a new annotation.

Given that currently the kubectl-gadget CLI creates the Trace with the
"operation=start" annotation and does not add any other annotation
later, it is useless to wait for the timeout if the Trace.Status
changes: State moves to "Completed"/"Started", or operationError/Warning
is set. In the end, such new Status will never change.