Implement fallback pod informer
This commit implements a mechanism that uses a pod informer as fallback
for other hooks. If a container is detected by the pod informer and it
hasn't been added to the container list. A warning message is printed
to informer the user that the main hook is not working properly.