cmd/utils: Improve CommonFlags
We don't have the Namespace field in the CommonFlags struct because
it's provided by KubernetesConfigFlags and we cannot add this flag
into the cobra command because it will be duplicated. To handle this
we have a GetNamespace() method to get the namespace, however it's
confusing to have all the parameters in one place but the namespace.

Now that we're using PersistentPreRunE to fill some fields in such
structure we can add the namespace in such structure and fill it
there.

This commit also adds a new "Labels map[string]string" fields to the
CommonFlags structure to avoid users having to parse them. The raw
representation is also kept in the "LabelsRaw" in case it's needed by
the user.