kubectl-gadget: Add check to ensure correct message is displayed
In a multi-node cluster, if a trace succeeds on at least one node and
the rest of the nodes (more than 1) report the same error message, the
function `printTraceFeedback()` will wrongly print "Failed to run the
gadget on all nodes" because all the messages it receives are the same.

The problem is that the function `printTraceFeedback()` does not know
the total number of nodes. With that information, it will be able to
conclude whether effectively all the nodes failed or not.

This commit also adds a unit test to cover this situation.