utils: trace: Wait for given initial state after creating the trace.
The goal of this commit is to give user a ready to use trace by waiting right
after its creation that it can be used.
For this, we need to wait on a particular state (often "Started").
Note that this is not needed for all gadgets, for example process-collector and
socket-collector do not need this.

Signed-off-by: Francis Laniel <flaniel@linux.microsoft.com>
Suggested-by: Jose Blanquicet <josebl@microsoft.com>
Fixes: a6bb14e8ff0f ("utils: trace: Add new API to interact with trace.")