utils: trace: Add handler for SIG* which cause program termination.
Previously to this commit, the following behavior was noticed:
./kubectl-gadget socket-collector -A | head -n0
kubectl get trace -A
NAMESPACE   NAME                      AGE
gadget      socket-collector-vtjxc    1s
The trace was not being deleted because kubectl-gadget was stopped receiving
SIGPIPE since it tried to write to a broken pipe.

In this commit, we install a handler for all signal which cause program
termination in RunTraceAndPrint*() to delete trace on reception of this signals.

Signed-off-by: Francis Laniel <flaniel@linux.microsoft.com>
Fixes: a6bb14e8ff0f ("utils: trace: Add new API to interact with trace.")