Include BTF files using BTFGen
BTFGen[0] allows to generate small BTF files with only the information
needed by a set of eBPF objects. This commit makes use of that tool
together with BTFHub[1] to create a set of BTF files needed by Inspektor
Gadget. These files are around 800KB and are included in the gadget
container.

[0]: https://github.com/kinvolk/btfgen
[1]: https://github.com/aquasecurity/btfhub