local-gadget: remove memlock rlimit
While gadgettracermanager increased the memlock rlimit, local-gadget was
not doing it and this leads to this error:

```
{"err":"failed to create BPF collection: program bpf_prog1: load
program: RLIMIT_MEMLOCK may be too low: operation not
permitted","notice":"failed to attach
tracer","node":"local","namespace":"default","pod":"shell01"}
```

This patch increases the memlock rlimit for local-gadget using the
provided function from cilium/ebpf and use the same code pattern in
gadgettracermanager.

Note that recent kernels don't require the memlock rlimit to be
increased when the process has CAP_IPC_LOCK (Linux >= 5.11). But it has
been found necessary for Linux 5.10.