collectors: Change operation from Start to Collect
Start is not an appropriate operation name for the gadgets that generate
a snapshot because they are not actually being started but just run. In
fact, after the Start operation they do not go in Started but Completed
and they do not even need to be stopped.

For these reasons, the operation name was changed from Start to Collect.