kubectl-gadget: Manage filtering by non-existent node
This commit moves into a common place for the BCC (previously not
covered) and non-BCC gadgets the node existence verification.

Notice this check will be removed when we will support the
addition/deletion of nodes.

Before this PR:
The gadget does not notify any error and remains blocked just doing
nothing.
---
$ ./kubectl-gadget opensnoop -A --node non-existent-node
^C
Terminating...
---

After this PR:
The gadget finishes immediately with the following error.
---
$ ./kubectl-gadget opensnoop -A --node non-existent-node
Error: invalid filter: node "non-existent-node" does not exist
---