seccomp: Record runc syscalls only after the profile is installed
The policies generated by the seccomp-advisor gadget are not enough
to get the pod to run successfully and, as a workaround, it is necessary
to set `allowPrivilegeEscalation` to `false` to make them work. However,
it does not always work.

This PR fixes a bug on the recording of the syscalls performed by
`runc`. It was previously relying on the fact that the `seccomp()` will
be called but it is not always true. Actually, it is false in most
cases.

Consider we want to record only the syscalls executed by `runc` after
the seccomp profile is actually installed. However, if we are running
the seccomp-advisor gadget, it's very probably that the pod does not
have a seccomp profile yet, so `seccomp()` will not be called.
Therefore, we decide to start recording from the
`prctl(PR_GET_PDEATHSIG)` call on. It is a safe place right before all
the `seccomp()` calls that will be always executed during the `runc`
initialisation.