New pkg: netnsenter
We used the same code pattern to enter a network namespace, execute some
code and return to the orginal network namespace. This patch factorizes
the code in a common package netnsenter.

This same code pattern will be used in a couple places more in the
tracepkt gadget, so it is time to factorize.