pkg: process-collector: Refactor the gadget.
Before this commit the following happens for, e.g., container name:
1. Container name was taken from gadget container golang.
2. It was written to eBPF map.
3. It was read from eBPF map and printed to seq file.
4. seq file for read from golang code.
5. The information were written to trace output.
6. The user get trace output using process-collector CLI.

For information we can only get through eBPF, like PID, this path is OK, but for
others, we should simplify it.
Thus, this commit refactored process-collector so it works like
socket-collector.
That is to say, data we can only get from eBPF are taken from there and the
others are handled directly in golang.

Signed-off-by: Francis Laniel <flaniel@linux.microsoft.com>