integration: Fix cleaning up commands
t.Fatal() is called when a test fails, it ends up calling runtime.Goexit()
to exit the current goroutime.

The Goexit documentation states that:

'''
Goexit terminates the goroutine that calls it. No other goroutine is affected.
Goexit runs all deferred calls before terminating the goroutine. Because Goexit
is not a panic, any recover calls in those deferred functions will return nil.
'''

The previous logic was using the same defer call to run the tests and
perform the cleanup, if a test failed then that function was interrupted
without doing the cleanup.

This commit uses a different defer call to perform the cleanup,
guaranteeing that it's always done.

Fixes: f6e6fdfd18e8 ("integration: Add start() and stop().")