cmd/kubectl-gadget: Implement tools-mode flag per gadget
The deploy command has a "tools-mode" flag that indicates which kind of
gadgets to use. The container entrypoint script has some logic to
determine the kind of tool to use when this is "auto". This logic
it not totally precise and some gadgets could not support the core
mode in some cases even if that logic indicates so.

In order to avoid the user having to redeploy IG, this commit adds a new
flag to each tool, so the user can choose the mode each time the tool is
run.

The previous existing "tools-mode" flag on the deploy command is renamed
to "default-tools-mode" as this can be useful to avoid having to pass
the mode each time a tool is executed.