integration: Fix cleaning up resources when tests complete
Yet another iteration on this one. The goal of TestMain() is to setup
Inspektor Gadget before running the test and tear it down after.
d061e9e4cd74 ("integration: Fix cleaning up commands") switched to use
defer in order to perform the clean up when the test fails. (It ends up
calling runtime.Goexit() which executes all defer calls). It worked fine
for that case but then when the test succeded it wasn't working fine
because os.Exit() doesn't call the defered functions.

This commit switches to an approach that mixes both of them, defer is
used in case the test fails and the cleanup functions are explictly
called before os.Exit().