kubectl-gadget: tolerate errors on some nodes
Ideally, there is one gadget pod on each worker node and it is in the
'running' phase. However, it could be that the gadget pod does not run
properly on one node. For example, Kubernetes could have a worker pool
with Windows and since the gadget container does not support Windows,
those containers are not running.

In this situation, it would still be useful to be able to use Inspektor
Gadget to inspect the other nodes. Currently, the kubectl-gadget CLI
fails immediately if one node does not work well.

This patch makes the CLI more tolerant of errors.