Fix localtime volume
/etc/localtime is normally a symlink to a regular file such as
/usr/share/zoneinfo/Etc/UTC. However, recent versions of Flatcar don't
have it anymore, and it could be created as a directory.

The gadget pod cannot control whether /etc/localtime on the host is a
directory or a file. This means it is not safe to have a host volume in
the gadget pod with the /etc/localtime -> /etc/localtime bind mount.

As a workaround, remove this host volume and instead let's have a
symlink /etc/localtime -> /host/etc/localtime in the pod. As opposed to
bind mounts, symlinks can be created regardless if the target is a
directory or a file.