Remove tools-mode flag
The tools-mode flag was intended to let the user choose the kind of
BCC tools to use: python-based/standard vs libbpf-based/core.

We have ported the tools to have their control plane in Golang
and we implemented a logic to automatically fallback to standard mode
when it's not possible to use CO:RE. For this reason this flag is not
useful anymore. In the future we could think about a way to specify the
tool mode (mainly for tests).