entrypoint.sh: Set current directory to a valid path
The previous logic was changing the directory to a path and then
removing it, which was causing issues when calling "getcwd" in BCC
python scripts.

This commit fixes the logic by using pushd/popd instead of plain cd, it
also changes the current directoy to / before executing
gadgettracermanager to be sure this problem doesn't happen again.