Check for fanotify support before selecting it
When using the hook-mode auto, we were only checking that runc was
present, but not that they system had fanotify capabilities. In systems
without `CONFIG_FANOTIFY` enabled, this would cause the Gadget Tracer
Manager to fail to start.

This change checks that `fanotify_init` returns successfully before
selecting it when in hook mode auto.

It also improves the error message when starting the tracer manager
fails, and avoids adding the fallback podinformer when the auto hook
mode selected podinformer, thus avoiding unnecessary warning messages.