Remove local-gadget from the default build target
The local-gadget target in our Makefile requires having CLang and
libbpf-dev installed locally, which are not considered required
dependencies for building Inspektor Gadget. The default build target
should only include targets that can be built with the listed tools.

This change removes local-gadget from "build", and creates an "all"
target that includes local-gadget