pkg/container-collection: external access to Kube API
go-client API has two ways to get the rest.Config:
1. from within the cluster (rest.InClusterConfig)
2. external access, with a kubeconfig

Some of Inspektor Gadget packages were hardcoding the 'in cluster'
config. This was fine for gadget-tracer-manager, which is meant to run
inside a pod. But this makes it more difficult to reuse our packages in
a different context.

This patch allows to optionally specify a different rest.Config.