local-gadget: Do not pin maps
local-gadget runs as a single process, hence it's not needed to pin
the eBPF maps. Instead, they can be passed as file descriptors
(or ebpf.Map). Pinning the maps is an issue in systems where bpffs is
not mounted by default (systemd <v238).

This commit modifies the gadgets to receive maps like ebpf.Maps objects.
Maps aren't pinned at all for local-gadget but they are still pinned in
the gadget tracer manager as the profile cpu gadgets still needs them.