Add a description field to all guides
The description written there appears in the overview page, making it a
lot more useful. These descriptions match what is displayed by the help
commands from the gadgets. In a few cases, the help message was modified
to be more consistent with the rest of the gadgets.

Also move the gadget guides to "20" and traceloop to "30", for sorting
purposes.