Review all guides
In this review, I checked that we are referring to the gadgets with the
new names instead of the old ones. I also reviewed for updating the
language to how we talk about the gadgets today (referring to them
individually rather than calling them Inspektor Gadget) and fixed some
typos.