gadgets: Remove Makefile files
The "ebpf-objects" target on gadget-container/Makefile forces to
recompile the ebpf objects, i.e, we're re-compiling all the eBPF objects
again when building the gadget container image, making it inefficient
and, in some cases, causing weird problems becuase the clang version
used inside the container is different than the one used to compiled the
shipped object.

We're now using go generate to invoke bpf2go in all the gadgets, hence
the makefiles in the bpf folders of all gadgets and the target mentioned
above are not needed anymore. Running "go generate ./..." on the root
folder will update all genereted files (invoking clang to compile the
ebpf programs.

This commit also fixes the usage of -D__TARGET when compiling eBPF code.

-D__TARGET has to be defined when compiling eBPF programs that include
<bpf/bpf_tracing.h> because the PT* macros defined there are dependent
on the architecture. Programs that don't include this file don't need
this to be defined.

The bpf2go command automatically defines this based on the -target used
[0], which means that we can completely remove this from our code.
However, we have to define the right architecture on the target flag
(instead of using bpf/bpfel/bpfeb) for programs that need this defined.

What it means for us is that some gadgets can use -target bpfel
(bpfeb on the future too) and use the same compiled object file for
multiple architectures, for other gadgets we need to compile for each
different architecture (only amd64 for the time being).

[0]: https://github.com/cilium/ebpf/blob/d17ebbefb05d7ebc1f1bac3b1c445369ff859f09/cmd/bpf2go/main.go#L295