pkg/gadgets: Improve standard tracer message when JSON unmarshalling failed.
With bcc tools, error message are not printed as JSON and some are print on
standard output.
For example, if the tool cannot find the kheaders, the following will be printed
on standard output:
Unable to find kernel headers. Try rebuilding kernel with CONFIG_IKHEADERS=m (module) or installing the kernel development package for your running kernel version.
Which would lead to the following error message being showed on user side:
failed to unmarshal event: invalid character 'U' looking for beginning of value

To fix this, we now print the whole line which failed to be marshalled:
failed to unmarshal event: invalid character 'U' looking for beginning of value
Unable to find kernel headers. Try rebuilding kernel with CONFIG_IKHEADERS=m (module) or installing the kernel development package for your running kernel version.

Signed-off-by: Francis Laniel <flaniel@linux.microsoft.com>